+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+
+### Improvements
+ - Add the "Last Duration" field in the "My Checks" page (#257)
+ - Add "last_duration" attribute to the Check API resource (#257)
+
+
 ## 1.9.0 - 2019-09-03
 
 ### Improvements